+++ examples/surreal_ql/database.py
@@ -130,6 +130,9 @@ with Database("http://127.0.0.1:8000", 'test', 'test', ('root', 'root')) as db:
     print(db.define_index("userAgeIndex", "user").columns("age"))
     # DEFINE INDEX userNameIndex ON TABLE user COLUMNS name SEARCH ANALYZER ascii BM25 HIGHLIGHTS;
     print(db.define_index("userNameIndex", "user").columns("name").search_analyzer("ascii").bm25().highlights())
+    # DEFINE INDEX userNameIndex ON TABLE user COLUMNS name SEARCH ANALYZER ascii BM25 1.2 0.75;
+    print(db.define_index("userNameIndex", "user").columns("name").search_analyzer("ascii").bm25(1.2, 0.75))
+
     # DEFINE INDEX mtree ON TABLE user FIELDS name MTREE DIMENSION 4 DIST EUCLIDEAN;
     print(db.define_index("mtree", "user").fields("name").mtree(4).distance_euclidean())
     # DEFINE INDEX hnsw ON TABLE user FIELDS name HNSW DIMENSION 4 DIST COSINE EFC 150 M 2;